Remove MessageExtensions
