--- src/Take.Blip.Builder/Content/MessageExtensions.cs
@@ -1,30 +1,0 @@
-﻿using Lime.Protocol;
-using Take.Blip.Client.Content;
-
-namespace Take.Blip.Builder.Content
-{
-    public static class MessageExtensions
-    {
-        public static Message CreateInputExirationTimeMessage(this Message message, string stateId) 
-        {
-            var idMessage = GetInputExirationTimeIdMessage(message);
-
-            return new Message(idMessage)
-            {
-                To = message.To,
-                Content = new InputExpiration()
-                { 
-                    Identity = message.From,
-                    StateId = stateId
-                }
-            };
-        }
-
-        public static string GetInputExirationTimeIdMessage(this Message message)
-        { 
-            return $"{message.From.ToIdentity()}-inputexpirationtime";
-        }
-
-
-    }
-}
